Handle return values of rsObject type
Bug: 25570907
Bug: 25777125

This made further improvements over a previous fix on return
values of rsObject type.

1) The temp variable created for a return expression is now set using
rsSetObject() rather than directly put in the initializer of the temp.
This way, the returned rsObject would have a sys ref count of at least
1 and won't become invalid on the return for the caller.

2) On the caller side, introduced another temp variable .rs.tmpXXX for
the result of any function call that returns an rsObject. This temp
is set in the initializer upon declaration, without using rsSetObject().
A destructor is inserted at the end of the enclosing scope for the temp.

3) Both temp variables above are now named with a unique suffix (that is
a unique integer, which starts from 0 and increases monotonically within
the function currently being analyzed and transformed.)

4) Fixed an issue in the previous fix, when there are multiple rsObjects
defined in a block, we will only introduce one temp variable for a
return statement. The previous CL would introduce one temp for each
rsObject, mistakenly.

Change-Id: Id3604f93b166089e3aca896d1c6c509b3ea19bcf