Fixed reference counting for function parameters
Bug: 25233132

For each function parameter that is object-typed or object-containing,
introduces a local variable and sets its value to the parameter.
This way, the sys ref counts of objects in the parameter are
incremented, preventing the objects being deleted if the parameter
is overwritten inside the function.

rsClearObj() is called on the rs_object-typed or -containing parameters
at the end of the function and at exiting points of the function.

This CL also fixes missed cases of returning structs that contain
rs objects, with an enhanced lit test to cover that.

Test: slang lit tests + RSTest and CTS on x86_64 emulator

Change-Id: Ic271c5e279b1f398ff233771b28d0acc5ac47c34
(cherry picked from commit 31d2ea3cb95fde60862e7032148c36f6edfbec5c)