Fix crashing in return that references rs obj
Bug: 38173182

Since slang adds rsClearObj() calls on local variables and
parameters of RS object types at exiting points, a return statement
may end up accessing nullptr if it references an RS object.

Previously, we handle the special case where the return value itself
is an RS object. This CL extends that treatment to the more general
case where the return value itself is not an RS object, but it
references an RS object.

The code transformation is the same as before for the special case,
i.e., to create a temporary variable for the return value, and
transform the return statement to return the temp instead.

Added a slang test for the general case.

Test: Slang Test
Test: CTS on Sailfish and X86_64 emulator

Change-Id: Idf29e648423290d87a428179d1bc38994525d413
(cherry picked from commit f05ee4b89abdaf166e6c42289b99a7b5cf7ee291)