Handle passed-via-stack vector arguments in Helper generation
On ARM 64-bit ABI(s), short-vector (or vector in Renderscript terms)
arguments could be passed via stack. When that happens, clang changes
the function's argument type to a pointer to a vector type.

The invokable function's type, subject to that potential modification,
is used by slang to generate the unpacking code (helper). However,
the reflected packing code in Java is generated per RS function type,
unaffected by this ABI-specific behavior.

So whever a short vector is passed via stack to an invokable, there
would be an inconsistency between the RS invokable function type and
LLVM function type.  This resulted in incorrect argument unpacking code
(helper) generated for invokables.

This CL fixes the inconsistency by packing the underlying element type
instead of the (pointer) type of the argument, as pointers are not valid
exportable types anyways. Helper code generation is also changed so that a
pointer is passed to the actual invokable function for the argument instead.

Bug: 36392046
Test: CTS on x86-64 emulator
Change-Id: I822d7f812471c3d3491571570573da726e3df2e0
(cherry picked from commit 48b49b4443305a4658cf2e3c65b5f56ea6667ed5)